+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Counter metrics now resolve measurements like other metric types, defaulting to 1 if measurement is not present.
+
 ## [1.5.0] - 2024-06-19
 
 ### Added